Create content.js:
content script that will be injected into the web page when the user right clicks on it. 
listen for the right-click event and then create a pop-up window at the position where the user clicked. 
pop-up window should contain a text box where the user can enter their note.
use the chrome.contextMenus API to create a context menu item that appears when the user right-clicks on the page. When the user clicks on this menu item, send a message to the content script to create the pop-up window.
save the page to a PDF,use the chrome.tabs.captureVisibleTab API to capture a screenshot of the current tab and then usejsPDF to convert the screenshot into a PDF.
use the chrome.downloads API to download the PDF to the user's computer when the save button is clicked.